// admin/src/app/services/api/models/SkillDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { SeniorityRatingEnum } from './SeniorityRatingEnum';

export type SkillDto = {
  id?: string;
  name?: string | null;
  yearsOfExperience?: number | null;
  seniorityRating?: SeniorityRatingEnum;
  personId?: string | null;
};